requirements.txt
python-dotenv
python-binance
tabulate
pyyaml